+++ src/components/common/TreeExplorer.vue
@@ -43,7 +43,7 @@ import { computed, provide, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 import TreeExplorerTreeNode from '@/components/common/TreeExplorerTreeNode.vue'
-import { useErrorHandling } from '@/composables/errorHooks'
+import { useErrorHandling } from '@/composables/useErrorHandling'
 import type {
   RenderedTreeExplorerNode,
   TreeExplorerNode
