+++ src/components/graph/GraphCanvas.vue
@@ -41,10 +41,10 @@ import TitleEditor from '@/components/graph/TitleEditor.vue'
 import NodeSearchboxPopover from '@/components/searchbox/NodeSearchBoxPopover.vue'
 import SideToolbar from '@/components/sidebar/SideToolbar.vue'
 import SecondRowWorkflowTabs from '@/components/topbar/SecondRowWorkflowTabs.vue'
-import { useCanvasDrop } from '@/composables/canvasDropHooks'
-import { useGlobalLitegraph } from '@/composables/litegraphHooks'
-import { useContextMenuTranslation } from '@/composables/translationHooks'
-import { useWorkflowPersistence } from '@/composables/workflowPersistenceHooks'
+import { useCanvasDrop } from '@/composables/useCanvasDrop'
+import { useContextMenuTranslation } from '@/composables/useContextMenuTranslation'
+import { useGlobalLitegraph } from '@/composables/useGlobalLitegraph'
+import { useWorkflowPersistence } from '@/composables/useWorkflowPersistence'
 import { CORE_SETTINGS } from '@/constants/coreSettings'
 import { i18n } from '@/i18n'
 import { api } from '@/scripts/api'
