+++ src/composables/useCanvasDrop.ts
@@ -2,7 +2,7 @@ import { LGraphNode } from '@comfyorg/litegraph'
 import { LiteGraph } from '@comfyorg/litegraph'
 import { Ref } from 'vue'
 
-import { usePragmaticDroppable } from '@/composables/dndHooks'
+import { usePragmaticDroppable } from '@/composables/usePragmaticDroppable'
 import { app as comfyApp } from '@/scripts/app'
 import { useLitegraphService } from '@/services/litegraphService'
 import { ComfyModelDef } from '@/stores/modelStore'
