+++ src/composables/useCopyToClipboard.ts
