+++ src/composables/useTreeExpansion.ts
