+++ src/services/extensionService.ts
@@ -1,4 +1,4 @@
-import { useErrorHandling } from '@/composables/errorHooks'
+import { useErrorHandling } from '@/composables/useErrorHandling'
 import { api } from '@/scripts/api'
 import { app } from '@/scripts/app'
 import { useCommandStore } from '@/stores/commandStore'
