+++ src/views/GraphView.vue
@@ -22,8 +22,8 @@ import UnloadWindowConfirmDialog from '@/components/dialog/UnloadWindowConfirmDi
 import GraphCanvas from '@/components/graph/GraphCanvas.vue'
 import GlobalToast from '@/components/toast/GlobalToast.vue'
 import TopMenubar from '@/components/topbar/TopMenubar.vue'
-import { useCoreCommands } from '@/composables/coreCommandHooks'
-import { useErrorHandling } from '@/composables/errorHooks'
+import { useCoreCommands } from '@/composables/useCoreCommands'
+import { useErrorHandling } from '@/composables/useErrorHandling'
 import { SERVER_CONFIG_ITEMS } from '@/constants/serverConfig'
 import { i18n } from '@/i18n'
 import { api } from '@/scripts/api'
